bots/requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.0
greenlet==3.0.2
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
numpy==1.26.2
pandas==2.1.4
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
yarl==1.9.4
